react component fixed
